+++ net.adamec.lib.common.core/logging/ILogger.CorrelationId.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.ComponentModel;
 
-namespace net.adamec.lib.common.logging
+namespace net.adamec.lib.common.core.logging
 {
     /// <summary>
     /// Logger interface - wrapper around the <see cref="NLog.ILogger"/> with some additional methods
     /// </summary>
-    internal partial interface ILogger 
+    public partial interface ILogger 
     {
          /// <summary>
         /// Writes the diagnostic message at the <c>Trace</c> level with correlation ID.