+++ net.adamec.lib.common.core/logging/ILogger.EventProperties.cs
@@ -2,12 +2,12 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 
-namespace net.adamec.lib.common.logging
+namespace net.adamec.lib.common.core.logging
 {
     /// <summary>
     /// Logger interface - wrapper around the <see cref="NLog.ILogger"/> with some additional methods
     /// </summary>
-    internal partial interface ILogger
+    public partial interface ILogger
     {
         /// <summary>
         /// Writes the diagnostic message at the <c>Trace</c> level.