+++ net.adamec.lib.common.core/logging/LoggerExt.EventProperties.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using NLog;
 
-namespace net.adamec.lib.common.logging
+namespace net.adamec.lib.common.core.logging
 {
     /// <summary>
     /// Extended logger implementing <see cref="ILogger"/>
     /// </summary>
-    internal partial class LoggerExt
+    public partial class LoggerExt
     {
         /// <summary>
         /// Writes the diagnostic message at the <c>Trace</c> level.