+++ net.adamec.lib.common/di/component/_SourceOnlyPackage.cs
@@ -6,8 +6,6 @@ namespace net.adamec.lib.common.di.component
     /// <NuProp.Description>RadCommons DI Component - allows to mark and inject the component using the class attributes (Source only package).</NuProp.Description>
     /// <NuProp.Tags>RadCommons source-only RadCommons-DI dependency-injection components</NuProp.Tags>
     /// <NuProp.Includes type="Folder" />
-    /// <NuProp.Using id="RadCommons.logging.CommonLogging" />
-    /// <NuProp.Using id="RadCommons.di.Component" />
     /// <NuProp.Needs id="Autofac" />
     [CompilerGenerated]
     // ReSharper disable once InconsistentNaming