+++ net.adamec.lib.common/di/postinit/_SourceOnlyPackage.cs
@@ -6,7 +6,6 @@ namespace net.adamec.lib.common.di.postinit
     /// <NuProp.Description>RadCommons DI PostInit - allows to run post init method of component (Source only package).</NuProp.Description>
     /// <NuProp.Tags>RadCommons source-only RadCommons-DI dependency-injection post-init</NuProp.Tags>
     /// <NuProp.Includes type="Folder" />
-    /// <NuProp.Using id="RadCommons.logging.CommonLogging" />
     /// <NuProp.Using id="RadCommons.di.Component" />
     /// <NuProp.Needs id="Autofac" />
     [CompilerGenerated]